src/services/Backend APIS/main-api/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pymysql==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.31.0
cryptography==41.0.3
